+++ lib/src/utils/youtube_player_flags.dart
@@ -34,7 +34,9 @@ class YoutubePlayerFlags {
   /// Forcing annotation to hide is a hacky way. Although this shouldn't be against Youtube TOS, the author doesn't guarantee
   /// and won't be responsible for any casualties regarding the YouTube TOS violation.
   ///
-  /// It's hidden by default on iOS. Changing this flag will have no effect on iOS.
+  /// Changing this flag will have no effect on iOS.
+  ///
+  /// Default is false.
   final bool forceHideAnnotation;
 
   /// Hides thumbnail if true.
@@ -69,7 +71,7 @@ class YoutubePlayerFlags {
     this.autoPlay = true,
     this.mute = false,
     this.isLive = false,
-    this.forceHideAnnotation = true,
+    this.forceHideAnnotation = false,
     this.hideThumbnail = false,
     this.disableDragSeek = false,
     this.enableCaption = true,